Fix MSVC narrowing warnings in DXF library
Make vector-size conversions explicit at DXF count boundaries so MSVC
does not warn about implicit size_t narrowing in 64-bit builds.

Write binary integer payload bytes through explicit casts and emit leader
vertex counts as integer DXF values instead of passing vector sizes through
the floating-point writer.

+++ libdxfrw/src/drw_entities.cpp
@@ -224,7 +224,7 @@ void DRW_Ellipse::toPolyline(DRW_Polyline *pol) {
     sinRot = sin(incAngle);
     incAngle = M_PI/64;
     curAngle = staparam;
-    int i = curAngle/incAngle;
+    int i = static_cast<int>(curAngle/incAngle);
     do {
         if (curAngle > endparam) {
             curAngle = endparam;
